# Have smaller fans