This increases the number of fans to get 4k large fans.
 
Before:
```
    2999k/3000k
3000000 bindings added in 10.507509s.
Finding 3000000 bindings.
    2999k/3000k
3000000 bindings found in 19.287851s (RW).
    2999k/3000k
3000000 bindings found in 22.156307s (RO).
```
After (30% speedup in RW mode):
```
Adding 3000000 bindings.
    2999k/3000k
3000000 bindings added in 10.506458s.
Finding 3000000 bindings.
    2999k/3000k
3000000 bindings found in 12.626557s (RW).
    2999k/3000k
3000000 bindings found in 18.202100s (RO).
```