@samoht 2019-09-16
I don't understand what this change is doing :-) Does it build fans with a fixed number of buckets?