@pascutto 2019-09-16
Not exactly, it builds fans so that each bucket contains ~4096 bytes. This way the fanout can scale to the number of elements in the `data` file.